[RTLIB_RPC] Raw performance counters support
To add raw performance counters append to the environment variable
BBQUE_RTLIB_OPTS, an option string having the following format:

"rN,LAB1-X,LAB2-Y"

where N is the number of performance counters, LABx is a text label
(consider max 5 chars) and X,Y... are the raw event code + unit mask of
the architecture, in the hexadecimal format.
For instance, if the event code is "E0h" and the unit mask is "7h", the
value to provide is "7E0".

If the raw performance counters mode is selected, all the default sets are
discarded.

NOTE: The label of the raw perfcounter is actually unused. In the future
it can be evaluated the possibility of adding it as a field of the
PerfEventAttr_t structure, and exploit it for a more readable dump.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>