[Worker] Add a first implementation
This is a base class which provides a bare minimum support to properly
initialize and track BarbequeRTRM support threads.
A worker is a thread which is started initially and run till the resource
manager is stopped. This class provides the necessary support to name
worker threads and properly terminate them when BBQ is terminated.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>