[PP] Switch to Worker usage
The PlatformProxy module is now converted to be a bbque::utils::Worker.
This exploits a code factorization and a better control on BBQ
threads setup and termination.

The Task thread is used to process platform notifications.
Once resources availability change, this thread should be run in
order to collect the new platform description.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>