[SC] Configuration parameters re-arrangement
The configuration parameters have been re-arranged such that we can have
more parameters, and for each parameter a per-resource value can be
provided. Actually, the MSL is the only general parameter provided.

Moreover, the penalties vectors in Congestion and Fairness have been
dimensioned to store a value per resource type. Therefore it is possible
to specify a penalty for each type of resource by adding
"penalty.<resource_type>" in the configuration file.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>