[RB] GetMask overloading
This adds two overloaded version of GetMask() member function.

One allows to retrieve a ResourceBitset object (from a UsagesMap) of a
specific resource type, given a resource scope, i.e., a pair type,ID that
allow to seek a position in a resource path, under which to look for the
target resource type.

The second version (called inside the first too), returns a ResourceBitset
related to a given type from a list of resource descriptors. The caller
can specify an application pointer and a resource state view to filter
the information related to a specific Application/EXC and/or resource state
view.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>