[CommandDispatcher 3/4] Add a simple and portable command line parser
This command line parser is a replacement of wordexpr for platform
not supporting that feature, e.g. Android. Of course it is much
more simple, indeed it support only up to 16 command line
parameters, which are expected to be space separated.

A single parameter could contain spaces if wrapped with single or
double quotes. Quoting with a symbol allows to use the other quote
simple within the parameter.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>