[AM] Bugfix AWM name reporting
When an application is not running is has not an assigned AWM, thus
in the log report, we must properly report that condition by printing
a '-' instead for the AWM name.

This patch properly initialize the AWM name for each application.
The default value will be replaced for the actual AWM name just for
applications with a current AWM.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>