[AM] Bugfix Current AWM reporting
When an application is not running is has not an assigned AWM, thus
in the log report, we must properly report that condition by printing
a '-' instead of the current AWM identifier.

This patch properly initialize the AWM identifier for each application.
The default value will be replaced for the current AWM name just for
applications which are already running.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>