[Tools_SignalAnalysis] Filter based Application instability
This patch adds a first implementation of an Application instability control
filter, which is based on the GoalGap variability EMA.
The basic idea is, a GG is forwarded only if the current GG variation is lower
than a tunable GG stability threshold.
This should block the issuing of too much variable GG requests, until the
application turns out to be more confident on the actual GG requirement.
For the time being, the GG threshold has a fix pre-configured value but we
expect to extend it to be correlated with the Reconfiguration Rate.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>