[Tools_SignalAnalysis] Add support for BC server
In order to speed-up simulation rpocess, we should avoid to spawn new
processes at run-time. Thus, using a single BC process for all the
computation should allows to get a significant speed-up.

This patch provides the required support to spawn a single BC process,
which takes code to execute from an input pipe and produce results into
an output pipe.
A convenient doComputation function is also available which could be used to
run the specified code and get results in the RESULTS output array.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>