[YM] Bugfix SyncLatency report on Sync calls case
When synchronous calls are in use, i.e. on the Android platform, the sync
latency reported by the RTLib during a PreChange was not processed by the
synchronization policy. This turned out on reading an always null sync
policy.
This patch fix this issue by ensuring to process all the possible return
code not only on asyn calls but also when synchronous calls are in use.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>