[SyncPol_SASB] Enforce sync ordering
The SASB policy defines a sub-ordering, within a sync step, which
considers the specific kind of decreasing (STEP2) or increasing (STEP3)
resources reconfiguration (i.e. MIGRATE, MIGREC, RECONF).

This patch ensure that this ordering is respected by properly updating
the DoSync call which is used by the synchronization steps of the
SynchronizationManager.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>