[YM] Bugfix platform synchronization ordering
The Platform sync, as well as all the other four sync protocol calls,
should check the synchronization policy before proceeding with the
tuning of platform constraints.

Without a policy check it could happens that the policy defined
synchronization ordering is not respected.
Such a check was missing and it has been fixed by this patch.

Moreover, given a certain sync queue, it could happen that the policy decide
that none of its applications should be synched. For example, all applications
in a RECONF queue are increasing their value, thus they should be synched after
the applications decreasing their value.

If, during a PreChange, no applications are going to be synched, than the
following stesp of the SyncP could be avoided.

NOTE: given a single queue, the sync policy could decide to do not sync any
app in a given step, i.e. DoSync return false.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>