[AM] Add support to release resources on Disable
When an EXC is disabled by the controlling application, resources
assigned to it should be immediately released to the system view.
This is required to have a system view which is always the most
up-to-date thus possibly better supporting an on-going scheduling
activity.

At the time being, applications could be disabled by the YM as well.
In that case resources release is not required. Thus the default
behavior of this new DisableEXC is to not release resources if not
explicitly required.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>