[RA] Optimize sync view creation
The creation of a new synchronization view requires the definition of a
tag which should be new for each sync run. A uint32_t counter is used to
define a new tag for each run, this patch slightly optimize the tag creation
by removing a check not required and using a smaller format string.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>