[YM] Move Reshuffling check on synchronization manager
Reshuffled EXC are those just changing their resources without actually
changing the AWM and/or their cluster. For these EXC, just the platform
layer should be updated, while the SyncP protocol with the RTLib could be
avoided.

This patch makes the YM to check for reshuffling conditions and avoid
SyncP messages for these EXC while still configuring the platform layer.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>