[YM] Add support to skip Sync Latency enforcing
This patch allows the Synchronization Manager (YM) to not enforce EXCs
reconfiguration to be aligned with the synchronization latency defined by
applications being reconfigured.

The synchronization latency is the time delay required by an application to
be in a synchronization point, i.e. complete the current run and being ready
to start a new one.

The estimation of the sync latency has been verified to be quite
sensible to run cycle time variation, thus producing too much
conservative reconfiguration points which noticeable affects the
behaviors of reconfigured applications.
Moreover, such a strict sync point enforcing is not mandatory in most
of the BBQ target platform, such as CGroups, PThreads and OpenCL
targets. Indeed, these platforms allows for a temporarily overlapping
of multiple applications on the same set of resources. The eventual
performance penalty is however to be considered negligible with
respect to the overheads introduced by the sync point enforcing.

By default, since this commit, the Sync Point enforcing is disabled
and should be explicitly enabled just when a resource contention
treat could be expected on a specific target platform or programming
paradigm.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>