[PP_Linux] Assign memory nodes based on BPL definition
The BarbequeRTRM Platform Layout (BPL) configuration file allows to
specify the memory node where a cluster is enabled to allocate memory.
This ID could be either a single value or a set/range according to the
same syntax used to define set/range of CPUs.

This patch exploits the Memory NODEs definition reported by the BPL
file to register the specified amount of memory for each one of the
specified nodes.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>