[SchedPol_YaMS] Initial version of COWS support
Cows is an extension of YaMS scheduler. The evaluation of the AWMs is
now splitted in two phases. In the first one, the evaluation procedes as
usual, computing the metrics and ordering the AWMs.  In the second
phase, the best binding of resources is selected. Each AWM is evaluated
on each binding domain through a metric that exploit information
extracted off-line from the applications. The goal is to promote the
simultaneous presence of cpu-bound and memory-bound applications on the
same binding domains, trying in the meanwhile to distribute resource
utilization evenly through the whole system.  Distributing the
applications in this way should improve system performance and avoid to
a certain degree hotspots and bottlenecks.

Signed-off-by: Simone Libutti <simone.libutti@gmail.com>