[RA] Add function GetTypesList()
With this patch the types of managed resources are tracked into a suitable
list. A member function has been added to the class in order to provide
such list to other BBQ components and plugins. The list is useful whenever
some scheduling contribution (e.g., Fairness) needs to iterate over all
the managed resource types to collect information.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>