[SchedPol_YaMS] Split scheduling contributions computation
This patch includes a huge refactoring of the YaMS policy aiming at
splitting the scheduling contributions computation between
binding-independent and binding-dependent.

Actually, to the former set belong Value, Reconfiguration and Fairness,
while to the latter one Congestion and Migration. Therefore the first set
is computed once for each pair (Application, AWM), while the second is
performed in the loop iterating over all the possible bindings.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>