[LinuxPP/AP/BbqueRPC] OpenCL device mapping
This introduces the OpenCL device mapping, if the support has been
enabled for compilation.

The LinuxPP instantiates an OpenCLProxy object as a sub-module, end
exploits its member functions to get the list of OpenCL platforms and
devices, and managing the mapping action. The latter, does nothing,
actually.

The Application Proxy checks if the resource set assigned includes
bindings on GPU resources. In that case, the ID of the GPU is sent to
the RTLib as a OpenCL device ID.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>