[SchedPol_YaMS] Multiple binding domains support
This commit introduces the possibility of specify multiple binding domains.
This in order to support the allocation of resources on heterogeneous
systems, e.g., featuring CPUs and GPUs.

In the configuration files, the option "binding.domain" can contain
multiple resource path separated by ','. BEWARE: No spaces are allowed.

The policy plugin allocate a scheduling contribution manager for each
binding domain, so that the final metrics computed takes into account
contributions computed on multiple binding domains.

Accordingly, the scheduling bindings temporarily stored in the AWM
descriptor are now referenced by a number (different from the previous
binding ID) computed as a function of the resource type of the binding
domain e the ID assigned.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>