[AWM] Multiple resource binding
Enable the possibility of performing resource binding on more than one
resource domain.

Member function BindResource() now returns a reference number through which
it is possible to reference the resource usage map already bound, on which
perform a binding related to another resource. The reference number now
replaces the previous usage of the "binding ID".

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>