[SchedPol_YaMS] Fix and improve multiple binding
With this patch the policy is now able to correctly perform multiple
bindings thanks to the exploitation of the scheduling binding reference
number provided by the WorkingMode objects.

Regarding the scheduling metrics, a "base value" is computed by considering
only the binding-independent contributions, afterwards the metrics value is
completed by adding the value returned by the recursive per binding domain
evaluation. More in detail, the value is the result of a mean between
the metrics value cumulated in the previous binding domain e the one compute
in the current one.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>