[YM] Revert to Sync Point Enforcing
The implementation of Sync Point Skipping is currently buggy and requires
some fixing in order to properly grant the reconfiguration of each
application which is skipping a sync. Thus, for the time being, let select
synchronization by default and mark this feature as EXPERIMENTAL.

TODO: fix the RTLib control cycle to ensure reconfiguration of applications
skipping a sync point.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>