[SchedPol_YaMS] Run-time change of metrics weights
This introduces a specific command for changing at run-time the values of
the scheduling contributions used for the metrics computation.

The number of arguments provided must be equal to the number of scheduling
contributions. To keep unchanged a weight, the character '-' can be used.

Syntax example: "bq.sp.yams.set_weights 20 5 10 -"

The order to consider for the contributions is AWMVALUE, RECONFIG, FAIRNESS,
MIGRATION, CONGESTION.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>