[AP] OpenCL device enforcement improvements
The ApplicationProxy gets the instance of the OpenCLProxy, such that in
case of GPU resource binding missing, we can send to the RTLib the correct
device ID related to the CPU. Previously, this was hard-coded to "2" for
early testing purposes. Now the AP exploits the device info functions
introduced in OCL_PP to retrieve the CPU ID (from the OpenCL Runtime
perspective).

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>