[Schedpol_YaMS] COWS: Robustness in initialization
In the CowsInit() the metrics contributions are now get by looping
on the array of contributions strings. This in order to make the
function implementation more flexible, whenever there is the need of
changing the set of contributions.
Morever each GetAttribute is safety-checked to ensure that a not null
pointer to the attribute object is returned (wrong Recipe case). In
that case the metrics value is set to 0 by default.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>