[RTLIB_RPC] Bugfix Sync Point Enforcing
If the sync-point is NOT enforced, the application is quite likely
to enter the GetWorkingMode _after_ the DoChange has been called which
was setting the AWM validity flag. Such a solution is not a problem
if the SyncPoint estimation is working properly since the application is
likely to enter the GetWorkingMode _before_ the DoChange is notified
by BBQ.

As a more general approach, an AWM is marked as "valid" only after
a GetAssignedWorkingMode has been performed. This allows to ensure
an execution of the onConfigure method by the AEM API.

To support such a solution a new AWM "assigned" flag has been added, which
is marked at DoChange notification time. Such a flag allows to get the
assigned AWM independently of the relative order among GWM and DoChange
entry time. Once the AWM has been assigned the AWM is marked as "valid" thus
allowing the application to go on with its processing loop.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>