[RTLIB] Add support for UNMANAGED execution mode
An application running in UNMANAGED mode is not controlled by the BarbequRTRM and
its automatically configured by the RTLib to run in AWM 0.
Once this support is enabled, an application could be executed un UNMANAGED mode
by setting the "U" flag in the BBQUE_RTLIB_OPTS environment variable.

This option is useful to run applications without BarbequeRTRM control, for example to
easily support Design Space Exploration (DSE) or the initial development stages of
a new application.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>