[OM] Changed fairness index formula
Since the app_var ranges from 0 to 0.25, the Fairness index has a
range limited in [1/1.25..1], which is not [0..1] as expected for
an evaluation index.

The index has thus been updated to solve this issue as well as to
enfore the following policy.
Given a certain number of READY applications, in a given priority
level, among two scheduling decisions which allocate a different
number of applications, the one with an higher number of scheduled
applications should always be preferred and thus it is expected to
get an higher FAIRNESS index.
Instead, if the number of scheduled applications is the same, the AWM
values variance is considered to be the main evaluation factor. Thus,
the higher the variance the lower the expected value of the FAIRNESS
index.

Resuming:
1. Scheduling choices with lower running/active apps ratio
are characterized by a lower index
2. Choices with the same running/active ratios are differentiated
using applications value variance as a metric

This new formulation:
   (running_count + (1 - 3 * awm_var)) / (1 + actives_count)
allows to enforce a monotonically increasing FAIRNESS index, with
range [0..1], which meets the aforementioned criteria.

Signed-off-by: Simone Libutti <simone.libutti@gmail.com>
Signed-off-by: Patrick Bellasi <derkling@gmail.com>