[OCL_PP] BBQUE_PLATFORM_NAME setting
The selection of the OpenCL platform to consider is performed by comparing
the sttribute "name" instead of "vendor" (BBQUE_PLATFORM_NAME).

A couple of macros have been introduced (AMD|INTEL_PLATFORM_NAME) as option
to choose for BBQUE_PLATFORM_NAME. The idea is to let the user to select
the reference platform through the Kconfig interface.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>