[RU] SetResourcesList overloading for filtered resources
The member function above has been overloaded to allows the assignment of
a filtered list of resources.
The filtering criteria is specified through a pair of additional parameters
('filter_rtype' and 'filter_mask'). The first one specifies the type of
resource to filter, while the second one the set of IDs to match for the
inclusion of the resource descriptor into the list.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>