[Logger 01/13] Switch loggers to be a BBQ utility
This patch is the first of a series which re-factors the BBQ logging
subsystem to be a statically linked module instead of a plugin.

Due to its fundamental nature it make no sens to have a run-time changed
logger, while it is quite common to select a logger at build time, perhaps
based on the target system and available logging libraries.

Such an implementation will allows to exploit Log4CPP for applications
coding as well.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>