[RA] UpdateResource: fix missing safety check
The pointer to ResourcePath returned by GetPath(...) was not checked.
It may happen that a not valid resource path is specified, leading to a
null pointer as return value. The bug has been discover through the
SetQuota/Reserve command, whenever a resource path like "sys0.cpu1.pe" was
specified. Indeed, in that case, the path does not reference a specific
resource descriptor. This condition raised a segmentation fault.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>