[Logger] Use default configuration file
The default configuration file provides a simple configuration of Log4CPP
which is suitable to be used by each application. Thus, if applications
will not specify a custom configuration, this one could be reference.

This patch setup the default configuration file for a Log4CPP logger.
This allows applications which does not specify a custom configuration,
to always get a default one. To the contrary, the BarbequeRTMR always
defines its own configuration file, thus the default value will be
overwritten.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>