[RTLIB-OCL] Forward clGetContextInfo at application startup
Once an application starts, ANY GPU device could be accessible thus allowing
the application to query the platform and setup a proper initialization
based on target platform capabilities.

This patch farwards all the clGetContextInfo calls when the R_ID_ANY device
ID is assiged, i.e. the default value when the application starts

Signed-off-by: Patrick Bellasi <derkling@gmail.com>