[PWRM 2/4] Refactor PowerManager base class
The PowerManager is now the only BBQ module which exposes a PM API to other
BBQ modules. The power management facilities for each and every specific
resource are now accessible via this single module which, by specifying
the resource path of interest, provides the required support to dispatch
the required call to the specific resource power manager.

Thus, for example, a call on a AMD GPU resource will be forwarded to the
AMDPowerManager module.

All the resource specific power manager could be defined and enabled at
compile time, based on the specific target configuration.
In case a resource specific power manager does not exposes a certain feature,
the main PowerManager module will notice that and does not forward the call
and it returns an PMResult::ERR_API_NOT_SUPPORTED error code.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>