[LinuxPP] Add simple support for CGroup Release Agent
The CGroup release agent is a user-space command which could be called by
the Linux kernel once a task exit from a control group.
This features could be exploited by BBQ to notice applications exiting
without a proper notification of the RTRM. In these cases the, a BBQ
provided release agent is called which in turn notify the RTRM by simply
sending an "unregister" command.

This patch add the support to process a "bq.pp.lnx.unregister" command.
For the time being, the command simply un-register the application

Signed-off-by: Patrick Bellasi <derkling@gmail.com>