[AM] Add support for EXC checking
This patch adds a couple of methods which allows to verify if the specified
EXC is still alive, i.e. the RTLib associated process is still in execution.

The optional parameter "release", by default enabled, allows to force a
destruction of the specified EXC in case it should be detected as a not
existing process.

These methods could be conveniently used to:
1. verify not-responding or timing-out EXCs being reconfigured
2. support a notify agent which allows to detect dead process

Signed-off-by: Patrick Bellasi <derkling@gmail.com>