[Tools_ReleaseAgent]
The CGroup release agent is a user-space command which could be called by
the Linux kernel once a task exit from a control group.
This features could be exploited by BBQ to notice applications exiting
without a proper notification of the RTRM.

This patch provides a BBQ release agent which is called by Linux kernel
to notify the RTRM about an application being terminated by simply sending
a "bq.pp.lnx.unregister" command.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>