[RTLIB_RPC] Remove not necessary debug statement
This statement is remove since it produce an unreadable output once compiled
in Debug mode. The issue is related to the value read from the CGroup
attribute, which produces a set of '\n' terminated lines.

A patch is provided but this debug code is kept commented to avoid a
awkward output.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>