[RTLIB_RPC] Add onMonitor profiling
This patch add support to collect statistics on AWM monitoring.
For each AWM, the onMonitor time required to execute the application provided
monitoring code is samples and accumulated.

NOTE: The profiling of these calls is done via the Notify API, since a direct
call to the onMonitor API is not exposed at plain API level. Thus, this
kind of profiling is available only when applications are integrated
via the AEM API.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>