[RTLIB_RPC] [DUMP 2/8] Setup statistics output file
Once enabled via 'f' flag, statistics are dumped in an output file which
reports the application UID and EXC. One file per EXC is produced.

An example filename is:
  stats_04321:bbque-testapp:exc_00
which reports statistics for the "bbque-testapp" application, insance with
PID 4321 and EXC 0.

These files are generated in the standard var directory, i.e. under:
  <out>/var/bbque

Signed-off-by: Patrick Bellasi <derkling@gmail.com>