[RTLIB_RPC] [DUMP 4/8] Switch to outfd handler
The outfd handler is properly configured to point to stderr or an opened
file, in case the dump on a file has been required via 'f' flag.

This patch updates all the calls related to statistics dumping, but the ones
related to the OpenCL support. These calls now make use of the aformentioned
outfd handler.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>