[RTLIB_RPC] Bugfix EXC UID formatting
The UID of an EXC is now built using up-to 13 chars for the application name,
but in case of a shorter name, only the provided chars will be used.

This name is used for example to setup the filename of the profiling output
file. Thus, this patch allows to avoid the creation of filenames with
spaces in between the name.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>