[RTLIB_CONF] Expose RTLib configuration to higer-level APIs
The RTLib configuration structure, which is defined and initilized by the
low-level API, is exposed to higher level APIs as an RTLib service.
The configuration structure is exposed as read-only and "mapped" to a
local "conf" member by the BbqueEXC base class. The AEM API exposes also
the configuration to applications via the BbqueEXC::Configuration method.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>