[RTLIB_CONF] Add DURATION enforcing support
This patch adds a new RTLib configuration option which allows to specify an
application running time upper bound in terms of
  a) number of cycles;
  b) seconds of processing time;

This option is available both in managed as well as in UNMANAGED mode,
but only by using the AEM API, i.e. for applications which extends the
BbqueEXC base class.

Once the specified upper limit is reached, the RTLib force a termination
of the control-loop, just as if the application should have reported
a RTLIB_EXC_WORKLOAD_NONE during the last onMonitor call.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>