[RTLIB_CGroups] Add CGrsoup enforcing configuration
A new configuration option has been added to the BBQUE_RTLIB_OPTS environment
variable, which allows to setup a CGroup for UNMANAGED applications.

The format string used to setup a CGroup is:
   "C <cpus> <cfs_period> <cfs_quota> <mems> <mem_bytes>"

CGroups enforcing is available only if:
1. the RTLib is build with CGroups support
2. the application is executed in UNMANAGED mode
3. the application is executed with ROOT privileges (necessary to setup a CGroup)

NOTE: for the time being, CGroups enforcing uses the same configuration
for all the EXC of an application. Thus, either:
a) the application uses a single EXC, or
b) the resources assigned should be sufficient to run all the EXCs
The solution a) is preferred.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>